feat(jenkins): did a vast transformation to move into using axios
We are now using axios instead of the jenkins api since it doesn't seem
to play too nicely with client-side JS applications. Also brought in a
docker compose for a re-deployable Jenkins instance wherever we go.
Might want to add in some general configurations into it so that it can
be spun up all at once.